test(validation): make the roster mirror comparison verbatim
The roster drift guard stripped both sides before comparing, so a
trailing-space edit to either the quoted table or the canonical one stayed
green under a test whose name claims a character-for-character mirror. It
now removes the literal comment indent instead. Verified it fires: adding
one trailing space to the quoted security row fails the guard, where the
stripped version passed.

Two cleanups on the same path. The evidence module carried an orphaned
comment describing template-fence extraction that moved to the boundaries
module in an earlier split. The boundaries module reached 506 lines after
running the formatter, so the canonical-source drift guards move to
test_git_hook_policy_adr_debate_mirrors.py along the same kind of seam as
the earlier split: the sibling asserts thresholds against text it writes,
this module reads documents the gate claims to mirror. The mutation
harness's _TESTS lists the third file, since the M8 killer moved with it.

_evidence_byte_count now names the residual a reviewer raised: an authored
U+FFFD loses three bytes. Undercounting can only block a log the committer
can lengthen; overcounting admits the invalid-byte bypass the measurement
exists to close.

Refs #5205
