+++ tests/mutation/test_mutate_adr_debate_evidence.py
@@ -66,6 +66,7 @@ _TARGET_REL = Path("scripts") / "validation" / "git_hook_policy.py"
 _TESTS = [
     "tests/validation/test_git_hook_policy_adr_debate_evidence.py",
     "tests/validation/test_git_hook_policy_adr_debate_boundaries.py",
+    "tests/validation/test_git_hook_policy_adr_debate_mirrors.py",
 ]
 
 _OUTCOME_DEAD = "DEAD"
@@ -75,7 +76,7 @@ _OUTCOME_DID_NOT_APPLY = "DID-NOT-APPLY"
 # Same ordering contract as the sibling harness: the outer cap MUST exceed the
 # inner one, or pytest-timeout interrupts inside subprocess.communicate and the
 # failure names no command (issue #5102). The inner suite here is
-# two files of 51 cases total, well under the sibling's ~943, so these caps
+# three files of 51 cases total, well under the sibling's ~943, so these caps
 # carry wide margin.
 _INNER_SUBPROCESS_TIMEOUT_SECONDS = 300
 _OUTER_TEST_TIMEOUT_SECONDS = 360
